test: add github issue test
